ignore locale_specific_texts.txt file

+++ create_dummy_from_en.sh
@@ -27,7 +27,7 @@ for FILE in ${LIST}; do
 	FILENAME_FLAT="${FILENAME_FLAT#*/}"
 	FILENAME_PREFIX="${FILENAME_FLAT%/*}"
 
-	if [[ ! -e "./input/$1/${FILENAME_FLAT}.txt" && "$FILENAME_PREFIX" != "base256" ]]; then
+	if [[ ! -e "./input/$1/${FILENAME_FLAT}.txt" && "${FILENAME_PREFIX}" != "base256"  && "${FILENAME_FLAT}" != "locale_specific_texts" ]]; then
 		echo "${FILENAME_FLAT} not found in input files, creating dummy ..."
 		mkdir -p "./input.new/$1/${FILENAME_FLAT%/*}"
 		echo -e "#TODO Please translate the following text:" > "./input.new/$1/${FILENAME_FLAT}.txt"
